Add common firebase service
