+++ src/common/services/firebase/index.js
@@ -1,0 +1,1 @@
+export * from './firebase.js';
