+++ src/index.js
@@ -1,9 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import './common/services/firebaseInit.js';
+import { init as initFirebase } from './common/services/firebase';
 import App from './common/components/App';
 import * as serviceWorker from './serviceWorker';
+
+initFirebase();
 
 ReactDOM.render(<App />, document.getElementById('root'));
