--- src/signIn/services/signIn.js
@@ -1,11 +1,0 @@
-import * as firebase from 'firebase/app';
-
-export function signIn(email, password, onError) {
-  // TODO: use async await syntax
-  firebase
-    .auth()
-    .signInWithEmailAndPassword(email, password)
-    .catch(function(error) {
-      onError(error.code);
-    });
-}
