--- src/signUp/services/signUp.js
@@ -1,11 +1,0 @@
-import * as firebase from 'firebase/app';
-
-export function signUp(email, password, onError) {
-  // TODO: use async await syntax
-  firebase
-    .auth()
-    .createUserWithEmailAndPassword(email, password)
-    .catch(function(error) {
-      onError(error.code);
-    });
-}
